Refactored call_hits_m8 and sort_m8_by_read_id to move it to blast.rs
