+++ Cargo.toml
@@ -17,9 +17,8 @@ rand = { version = "0.9.2", features = ["std_rng"] }
 rand_distr = "0.5.1"
 futures = "0.3.31"
 sysinfo = "0.37.0"
-fxhash = "0.2.1"
 num_cpus = "1.16"
-bio = "2.3.0"
+bio = "4.0.0"
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.140"
 plotters = "0.3.7"
@@ -31,7 +30,6 @@ needletail = "0.6.3"
 uuid = { version = "1.18.1", features = ["v4"] }
 twox-hash = "1.6"
 rand_core = "0.6.4"
-sled = "0.34"
 csv = "1.3"
 lexical = "6.1.1"
 fst = "0.4.7"
@@ -43,7 +41,6 @@ bytes = "1.11.1"
 dashmap = "6.1.0"
 regex = "1.12.2"
 once_cell = "1.21.3"
-async-stream = "0.3.6"
 nix = { version = "0.31.2", features = ["fs"] }
 which = "8.0.2"
 proptest = "1.11.0"
